GO-2579 Simplify creator
