+++ core/block/object/objectcreator/creator.go
@@ -3,25 +3,16 @@ package objectcreator
 import (
 	"context"
 	"fmt"
-	"strings"
-	"time"
 
 	"github.com/anyproto/any-sync/app"
-	"github.com/globalsign/mgo/bson"
 	"github.com/gogo/protobuf/types"
-	"golang.org/x/exp/slices"
 
 	"github.com/anyproto/anytype-heart/core/block/bookmark"
-	"github.com/anyproto/anytype-heart/core/block/editor/objecttype"
-	"github.com/anyproto/anytype-heart/core/block/editor/smartblock"
 	"github.com/anyproto/anytype-heart/core/block/editor/state"
-	"github.com/anyproto/anytype-heart/core/block/object/objectcache"
 	"github.com/anyproto/anytype-heart/core/domain"
-	"github.com/anyproto/anytype-heart/metrics"
 	"github.com/anyproto/anytype-heart/pb"
 	"github.com/anyproto/anytype-heart/pkg/lib/bundle"
 	coresb "github.com/anyproto/anytype-heart/pkg/lib/core/smartblock"
-	"github.com/anyproto/anytype-heart/pkg/lib/localstore/addr"
 	"github.com/anyproto/anytype-heart/pkg/lib/localstore/objectstore"
 	"github.com/anyproto/anytype-heart/pkg/lib/logging"
 	"github.com/anyproto/anytype-heart/pkg/lib/pb/model"
@@ -30,18 +21,24 @@ import (
 	"github.com/anyproto/anytype-heart/util/pbtypes"
 )
 
+type (
+	CollectionService interface {
+		CreateCollection(details *types.Struct, flags []*model.InternalFlag) (coresb.SmartBlockType, *types.Struct, *state.State, error)
+	}
+
+	TemplateService interface {
+		CreateTemplateStateWithDetails(templateId string, details *types.Struct) (st *state.State, err error)
+		TemplateCloneInSpace(space space.Space, id string) (templateId string, err error)
+	}
+)
+
+const CName = "objectCreator"
+
 var log = logging.Logger("object-service")
-
-type eventKey int
-
-const eventCreate eventKey = 0
 
 type Service interface {
 	CreateObject(ctx context.Context, spaceID string, req CreateObjectRequest) (id string, details *types.Struct, err error)
-	CreateObjectInSpace(ctx context.Context, space space.Space, req CreateObjectRequest) (id string, details *types.Struct, err error)
-	CreateObjectUsingObjectUniqueTypeKey(ctx context.Context, spaceID string, objectUniqueTypeKey string, req CreateObjectRequest) (id string, details *types.Struct, err error)
 	CreateSmartBlockFromState(ctx context.Context, spaceID string, objectTypeKeys []domain.TypeKey, createState *state.State) (id string, newDetails *types.Struct, err error)
-	CreateSmartBlockFromStateInSpace(ctx context.Context, space space.Space, objectTypeKeys []domain.TypeKey, createState *state.State) (id string, newDetails *types.Struct, err error)
 
 	InstallBundledObjects(ctx context.Context, space space.Space, sourceObjectIds []string, isNewSpace bool) (ids []string, objects []*types.Struct, err error)
 	app.Component
@@ -54,15 +51,6 @@ type service struct {
 	app               *app.App
 	spaceService      space.Service
 	templateService   TemplateService
-}
-
-type CollectionService interface {
-	CreateCollection(details *types.Struct, flags []*model.InternalFlag) (coresb.SmartBlockType, *types.Struct, *state.State, error)
-}
-
-type TemplateService interface {
-	CreateTemplateStateWithDetails(templateId string, details *types.Struct) (st *state.State, err error)
-	TemplateCloneInSpace(space space.Space, id string) (templateId string, err error)
 }
 
 func NewCreator() Service {
@@ -79,89 +67,8 @@ func (s *service) Init(a *app.App) (err error) {
 	return nil
 }
 
-const CName = "objectCreator"
-
 func (s *service) Name() (name string) {
 	return CName
-}
-
-// CreateSmartBlockFromState create new object from the provided `createState` and `details`.
-// If you pass `details` into the function, it will automatically add missing relationLinks and override the details from the `createState`
-// It will return error if some of the relation keys in `details` not installed in the workspace.
-func (s *service) CreateSmartBlockFromState(
-	ctx context.Context, spaceID string, objectTypeKeys []domain.TypeKey, createState *state.State,
-) (id string, newDetails *types.Struct, err error) {
-	spc, err := s.spaceService.Get(ctx, spaceID)
-	if err != nil {
-		return "", nil, err
-	}
-	return s.CreateSmartBlockFromStateInSpace(ctx, spc, objectTypeKeys, createState)
-}
-
-func (s *service) CreateSmartBlockFromStateInSpace(
-	ctx context.Context, spc space.Space, objectTypeKeys []domain.TypeKey, createState *state.State,
-) (id string, newDetails *types.Struct, err error) {
-	if createState == nil {
-		createState = state.NewDoc("", nil).(*state.State)
-	}
-	startTime := time.Now()
-	// priority:
-	// 1. details
-	// 2. createState
-	// 3. createState details
-	// 4. default object type by smartblock type
-	if len(objectTypeKeys) == 0 {
-		objectTypeKeys = []domain.TypeKey{bundle.TypeKeyPage}
-	}
-	sbType := objectTypeKeysToSmartBlockType(objectTypeKeys)
-
-	createState.SetDetailAndBundledRelation(bundle.RelationKeySpaceId, pbtypes.String(spc.Id()))
-
-	ev := &metrics.CreateObjectEvent{
-		SetDetailsMs: time.Since(startTime).Milliseconds(),
-	}
-
-	ctx = context.WithValue(ctx, eventCreate, ev)
-	initFunc := func(id string) *smartblock.InitContext {
-		createState.SetRootId(id)
-		return &smartblock.InitContext{
-			Ctx:            ctx,
-			ObjectTypeKeys: objectTypeKeys,
-			State:          createState,
-			RelationKeys:   generateRelationKeysFromState(createState),
-			SpaceID:        spc.Id(),
-		}
-	}
-
-	sb, err := createSmartBlock(ctx, spc, initFunc, createState, sbType)
-	if err != nil {
-		return "", nil, err
-	}
-
-	newDetails = sb.CombinedDetails()
-	id = sb.Id()
-
-	if sbType == coresb.SmartBlockTypeObjectType && pbtypes.GetInt64(newDetails, bundle.RelationKeyLastUsedDate.String()) == 0 {
-		objecttype.UpdateLastUsedDate(spc, s.objectStore, domain.TypeKey(
-			strings.TrimPrefix(pbtypes.GetString(newDetails, bundle.RelationKeyUniqueKey.String()), addr.ObjectTypeKeyToIdPrefix)),
-		)
-	} else if pbtypes.GetInt64(newDetails, bundle.RelationKeyOrigin.String()) == int64(model.ObjectOrigin_none) {
-		objecttype.UpdateLastUsedDate(spc, s.objectStore, objectTypeKeys[0])
-	}
-
-	ev.SmartblockCreateMs = time.Since(startTime).Milliseconds() - ev.SetDetailsMs - ev.WorkspaceCreateMs - ev.GetWorkspaceBlockWaitMs
-	ev.SmartblockType = int(sbType)
-	ev.ObjectId = id
-	metrics.SharedClient.RecordEvent(*ev)
-	return id, newDetails, nil
-}
-
-func getUniqueKeyOrGenerate(sbType coresb.SmartBlockType, details *types.Struct) (domain.UniqueKey, error) {
-	uniqueKey := pbtypes.GetString(details, bundle.RelationKeyUniqueKey.String())
-	if uniqueKey == "" {
-		return domain.NewUniqueKey(sbType, bson.NewObjectId().Hex())
-	}
-	return domain.UnmarshalUniqueKey(uniqueKey)
 }
 
 // TODO Add validate method
@@ -170,18 +77,30 @@ type CreateObjectRequest struct {
 	InternalFlags []*model.InternalFlag
 	TemplateId    string
 	ObjectTypeKey domain.TypeKey
+	UniqueKey     string
 }
 
+// CreateObject is high-level method for creating new objects
 func (s *service) CreateObject(ctx context.Context, spaceID string, req CreateObjectRequest) (id string, details *types.Struct, err error) {
 	space, err := s.spaceService.Get(ctx, spaceID)
 	if err != nil {
 		return "", nil, fmt.Errorf("get space: %w", err)
 	}
-	return s.CreateObjectInSpace(ctx, space, req)
+
+	if req.ObjectTypeKey.String() == "" {
+		objectTypeKey, err := domain.GetTypeKeyFromRawUniqueKey(req.UniqueKey)
+		if err != nil {
+			return "", nil, fmt.Errorf("get type key from raw unique key: %w", err)
+		}
+		req.ObjectTypeKey = objectTypeKey
+	}
+
+	return s.createObjectInSpace(ctx, space, req)
 }
 
-// CreateObjectInSpace is high-level method for creating new objects
-func (s *service) CreateObjectInSpace(ctx context.Context, space space.Space, req CreateObjectRequest) (id string, details *types.Struct, err error) {
+func (s *service) createObjectInSpace(
+	ctx context.Context, space space.Space, req CreateObjectRequest,
+) (id string, details *types.Struct, err error) {
 	details = req.Details
 	if details.GetFields() == nil {
 		details = &types.Struct{Fields: map[string]*types.Value{}}
@@ -190,12 +109,12 @@ func (s *service) CreateObjectInSpace(ctx context.Context, space space.Space, re
 
 	switch req.ObjectTypeKey {
 	case bundle.TypeKeyBookmark:
-		return s.ObjectCreateBookmark(ctx, space.Id(), &pb.RpcObjectCreateBookmarkRequest{
+		return s.createBookmark(ctx, space.Id(), &pb.RpcObjectCreateBookmarkRequest{
 			Details: details,
 		})
 	case bundle.TypeKeySet:
 		details.Fields[bundle.RelationKeyLayout.String()] = pbtypes.Float64(float64(model.ObjectType_set))
-		return s.CreateSet(ctx, space, &pb.RpcObjectCreateSetRequest{
+		return s.createSet(ctx, space, &pb.RpcObjectCreateSetRequest{
 			Details:       details,
 			InternalFlags: req.InternalFlags,
 			Source:        pbtypes.GetStringList(details, bundle.RelationKeySetOf.String()),
@@ -207,7 +126,7 @@ func (s *service) CreateObjectInSpace(ctx context.Context, space space.Space, re
 		if err != nil {
 			return "", nil, err
 		}
-		return s.CreateSmartBlockFromStateInSpace(ctx, space, []domain.TypeKey{bundle.TypeKeyCollection}, st)
+		return s.createSmartBlockFromStateInSpace(ctx, space, []domain.TypeKey{bundle.TypeKeyCollection}, st)
 	case bundle.TypeKeyObjectType:
 		return s.createObjectType(ctx, space, details)
 	case bundle.TypeKeyRelation:
@@ -216,19 +135,10 @@ func (s *service) CreateObjectInSpace(ctx context.Context, space space.Space, re
 		return s.createRelationOption(ctx, space, details)
 	}
 
-	return s.createSmartBlockFromTemplate(ctx, space, []domain.TypeKey{req.ObjectTypeKey}, details, req.TemplateId)
+	return s.createObjectFromTemplate(ctx, space, []domain.TypeKey{req.ObjectTypeKey}, details, req.TemplateId)
 }
 
-func (s *service) CreateObjectUsingObjectUniqueTypeKey(ctx context.Context, spaceID string, objectUniqueTypeKey string, req CreateObjectRequest) (id string, details *types.Struct, err error) {
-	objectTypeKey, err := domain.GetTypeKeyFromRawUniqueKey(objectUniqueTypeKey)
-	if err != nil {
-		return "", nil, fmt.Errorf("get type key from raw unique key: %w", err)
-	}
-	req.ObjectTypeKey = objectTypeKey
-	return s.CreateObject(ctx, spaceID, req)
-}
-
-func (s *service) createSmartBlockFromTemplate(
+func (s *service) createObjectFromTemplate(
 	ctx context.Context,
 	space space.Space,
 	objectTypeKeys []domain.TypeKey,
@@ -239,57 +149,5 @@ func (s *service) createSmartBlockFromTemplate(
 	if err != nil {
 		return
 	}
-	return s.CreateSmartBlockFromStateInSpace(ctx, space, objectTypeKeys, createState)
-}
-
-func objectTypeKeysToSmartBlockType(typeKeys []domain.TypeKey) coresb.SmartBlockType {
-	// TODO Add validation for types that user can't create
-
-	if slices.Contains(typeKeys, bundle.TypeKeyTemplate) {
-		return coresb.SmartBlockTypeTemplate
-	}
-	typeKey := typeKeys[0]
-
-	switch typeKey {
-	case bundle.TypeKeyObjectType:
-		return coresb.SmartBlockTypeObjectType
-	case bundle.TypeKeyRelation:
-		return coresb.SmartBlockTypeRelation
-	case bundle.TypeKeyRelationOption:
-		return coresb.SmartBlockTypeRelationOption
-	default:
-		return coresb.SmartBlockTypePage
-	}
-}
-
-func createSmartBlock(
-	ctx context.Context, spc space.Space, initFunc objectcache.InitFunc, st *state.State, sbType coresb.SmartBlockType,
-) (smartblock.SmartBlock, error) {
-	if uKey := st.UniqueKeyInternal(); uKey != "" {
-		uk, err := domain.NewUniqueKey(sbType, uKey)
-		if err != nil {
-			return nil, err
-		}
-		return spc.DeriveTreeObject(ctx, objectcache.TreeDerivationParams{
-			Key:      uk,
-			InitFunc: initFunc,
-		})
-	}
-	return spc.CreateTreeObject(ctx, objectcache.TreeCreationParams{
-		Time:           time.Now(),
-		SmartblockType: sbType,
-		InitFunc:       initFunc,
-	})
-}
-
-func generateRelationKeysFromState(st *state.State) (relationKeys []string) {
-	if st == nil {
-		return
-	}
-	details := st.CombinedDetails().GetFields()
-	relationKeys = make([]string, 0, len(details))
-	for k := range details {
-		relationKeys = append(relationKeys, k)
-	}
-	return
+	return s.createSmartBlockFromStateInSpace(ctx, space, objectTypeKeys, createState)
 }
