+++ core/block/object/objectcreator/relation.go
@@ -58,5 +58,5 @@ func (s *service) createRelation(ctx context.Context, space space.Space, details
 
 	createState := state.NewDocWithUniqueKey("", nil, uniqueKey).(*state.State)
 	createState.SetDetails(object)
-	return s.CreateSmartBlockFromStateInSpace(ctx, space, []domain.TypeKey{bundle.TypeKeyRelation}, createState)
+	return s.createSmartBlockFromStateInSpace(ctx, space, []domain.TypeKey{bundle.TypeKeyRelation}, createState)
 }
