+++ core/block/object/objectcreator/relation_option.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/globalsign/mgo/bson"
 	"github.com/gogo/protobuf/types"
 
 	"github.com/anyproto/anytype-heart/core/block/editor/state"
@@ -39,5 +40,13 @@ func (s *service) createRelationOption(ctx context.Context, space space.Space, d
 
 	createState := state.NewDocWithUniqueKey("", nil, uniqueKey).(*state.State)
 	createState.SetDetails(object)
-	return s.CreateSmartBlockFromStateInSpace(ctx, space, []domain.TypeKey{bundle.TypeKeyRelationOption}, createState)
+	return s.createSmartBlockFromStateInSpace(ctx, space, []domain.TypeKey{bundle.TypeKeyRelationOption}, createState)
+}
+
+func getUniqueKeyOrGenerate(sbType coresb.SmartBlockType, details *types.Struct) (domain.UniqueKey, error) {
+	uniqueKey := pbtypes.GetString(details, bundle.RelationKeyUniqueKey.String())
+	if uniqueKey == "" {
+		return domain.NewUniqueKey(sbType, bson.NewObjectId().Hex())
+	}
+	return domain.UnmarshalUniqueKey(uniqueKey)
 }
